Add an abstract class
